Documentation, renames and general refactoring.
