+++ pkg/kev/volumes.go
@@ -1,0 +1,31 @@
+/**
+ * Copyright 2020 Appvia Ltd <info@appvia.io>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package kev
+
+// condenseLabels returns a copy of the VolumeConfig with only condensed base volume labels
+func (vc VolumeConfig) condenseLabels(labels []string) VolumeConfig {
+	for key := range vc.Labels {
+		if !contains(labels, key) {
+			delete(vc.Labels, key)
+		}
+	}
+
+	return VolumeConfig{
+		Name:   vc.Name,
+		Labels: vc.Labels,
+	}
+}
